2017-01-15-heating-daemon-rewrite
WIP on getting the rulesD working, mainly the ControlsCurrentState.pm
module with the methods for getting and setting controls, pushing to the DB.

The methods will be able to test for the standardised types of state.
(mentioned below

The ControlsCurrentState.pm will also be usuable by the StatusD.

A minor schema update with a controls.config_json field.

Also constants that state that :

On  = true  = Open   = Pingable     = 1
Off = false = Closed = not-pingable = 0

These are meaningful in the context of different control types.