2017-01-15-heating-daemon-rewrite
added to one-wire-controls the 'display_webui_float_control'
to allow association between a one-wire-therm-control and a
webui-var-float-control.

This will be so the webui can display different colours if the control is
in the desired range.

sub moved from Khaospy::Message and put into Khaospy::Conf::Controls
to validate_control_state_action. This still requires more work.
The end goal is for this to validate that a state or action is valid for the control.

minor tidy up of the Khaospy::OneWireThermometer (daemon) code.