New Khaospy::Conf::Global module
settings that are currently constants, but that will be configurable
will migrate from Khaospy::Constants into Khaospy::Conf::Global.

The default value will be hardcoded in Khaospy::Conf::Global,
and they will be removed from Constants.

Bit of tidy up on the Rooms Status (rooms.js and tt), and simplification.